requirements.txt
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
